services/call_to_action/requirements.txt
pycld3==0.22
pymorphy2==0.9.1
spacy==3.5.4
https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.5.0/ru_core_news_sm-3.5.0-py3-none-any.whl
urllib3==1.26.14
